Back-End/requirements.txt
python-dotenv
flask_cors
pymysql
flask
